Fix costs and profits parsing

+++ src/onedimensional/instance_builder.cpp
@@ -298,7 +298,7 @@ void InstanceBuilder::read_bin_types(
             if (labels[i] == "X") {
                 x = (Length)std::stol(line[i]);
             } else if (labels[i] == "COST") {
-                cost = (Profit)std::stol(line[i]);
+                cost = (Profit)std::stod(line[i]);
             } else if (labels[i] == "COPIES") {
                 copies = (BinPos)std::stol(line[i]);
             } else if (labels[i] == "COPIES_MIN") {
